pulling out sample set class separate from main
